Refactor messages
