--- AzureCommunicationUI/sdk/AzureCommunicationUIChat/Sources/Presentation/SwiftUI/Chat/ChatViewComponents/Message/HtmlMessageViewModel.swift
@@ -1,9 +1,0 @@
-//
-//  Copyright (c) Microsoft Corporation. All rights reserved.
-//  Licensed under the MIT License.
-//
-
-import Foundation
-
-class HtmlMessageViewModel: MessageViewModel {
-}
